introduce PartitionKey/RowKey attributes
The PartitionKey and RowKey attributes are to inform the framework that
a given string-valued property of an object is to be used as the
partition or row key when storing/retrieving that object from table
storage.

We extend serialization/deserialization of GenericEntity to account
for PartitionKey and RowKey attributes.

We also add method overloads to TableStorageProvider, and also to
ITableContext and implementors, for adding an item without supplying
a partition/row key. These methods assume the PartitionKey and RowKey
attributes decorate suitable properties of the item's type.