fix/centralize property serialization logic
Logic for deciding to serialize a property was duplicated
(and inconsistent) in both MemoryTableContext and AzureTableContext.
We centralize it as an extension method in PropertyInfoExtensions.cs.

We also change the logic to search a property's inheritence chain
for DontSerialize attributes. If a class flags a property as
DontSerialize, subclasses should not have to repeat the property.
THIS IS A BREAKING CHANGE!

If any code written against DataModel.Core depends on the above
behavior, that code will be broken.